requirements.txt
pandas==2.2.3
numpy
python-dotenv==1.0.1
llama-index==0.12.17
llama-index-core==0.12.17
llama-index-llms-openai==0.3.19
llama-index-embeddings-openai==0.3.1
openai==1.63.0
openpyxl==3.1.5
flask==3.0.3
werkzeug==3.0.1